move add book modal to own folder
