+++ frontend/src/my-books/MyBooks.tsx
@@ -18,7 +18,7 @@ If not, see <https://www.gnu.org/licenses/>.
 import React, { Component, ReactElement } from 'react';
 import { NavBar } from '../shared/navigation/NavBar';
 import Button from '@material-ui/core/Button';
-import AddBookModal from './AddBookModal';
+import AddBookModal from '../add-book-modal/AddBookModal';
 import { Layout } from '../shared/components/Layout';
 import { Genres } from '../shared/types/Genres';
 import { Book } from '../shared/types/Book';
